+++ src/Symfony/Component/Validator/Constraints/GroupSequence.php
@@ -28,9 +28,7 @@ namespace Symfony\Component\Validator\Constraints;
  *
  * When adding metadata to a class, you can override the "Default" group of
  * that class with a group sequence:
- *     /**
- *      * @GroupSequence({"Address", "Strict"})
- *      *\/
+ *     #[GroupSequence(['Address', 'Strict'])]
  *     class Address
  *     {
  *         // ...
@@ -45,9 +43,6 @@ namespace Symfony\Component\Validator\Constraints;
  * with an overridden default group, pass the class name as group name instead:
  *
  *     $validator->validate($address, null, "Address")
- *
- * @Annotation
- * @Target({"CLASS", "ANNOTATION"})
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */