+++ src/Symfony/Component/Validator/Constraints/NotBlank.php
@@ -15,9 +15,6 @@ use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Exception\InvalidArgumentException;
 
 /**
- * @Annotation
- * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
- *
  * @author Bernhard Schussek <bschussek@gmail.com>
  * @author Kévin Dunglas <dunglas@gmail.com>
  */
@@ -32,6 +29,7 @@ class NotBlank extends Constraint
 
     public $message = 'This value should not be blank.';
     public $allowNull = false;
+    /** @var callable|null */
     public $normalizer;
 
     public function __construct(array $options = null, string $message = null, bool $allowNull = null, callable $normalizer = null, array $groups = null, mixed $payload = null)