+++ src/Symfony/Component/Validator/Constraints/Regex.php
@@ -15,9 +15,6 @@ use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Exception\InvalidArgumentException;
 
 /**
- * @Annotation
- * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
- *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
 #[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::TARGET_METHOD | \Attribute::IS_REPEATABLE)]
@@ -33,6 +30,7 @@ class Regex extends Constraint
     public $pattern;
     public $htmlPattern;
     public $match = true;
+    /** @var callable|null */
     public $normalizer;
 
     public function __construct(