+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -5,12 +5,16 @@ CHANGELOG
 ---
 
  * Add argument `$routeParameters` to `UrlMatcher::handleRouteRequirements()`
+ * Remove Doctrine annotations support in favor of native attributes
+ * Change the constructor signature of `AnnotationClassLoader` to `__construct(?string $env = null)`, passing an annotation reader as first argument is not supported anymore
 
 6.4
 ---
 
  * Add FQCN and FQCN::method aliases for routes loaded from attributes/annotations when applicable
  * Add native return type to `AnnotationClassLoader::setResolver()`
+ * Deprecate Doctrine annotations support in favor of native attributes
+ * Change the constructor signature of `AnnotationClassLoader` to `__construct(?string $env = null)`, passing an annotation reader as first argument is deprecated
 
 6.2
 ---